feat: relocate Glow Sync Hub entry to Collection Hub with enhanced visuals
This commit moves the Glow Sync Hub access point from the vanity landing screen to the collection hub. The transition includes the implementation of a shimmer animation effect and UI refinements to make the entry point more prominent within the collection interface.

- **Collection Hub Integration**:
    - Added a "Glow Sync Hub" card to `CollectionHubScreen.kt` that navigates to the starter pack route.
    - Implemented a shimmer animation using `rememberInfiniteTransition` and a linear gradient brush to create a dynamic "glow" effect.
    - Styled the hub entry with a deep plum background (`0xFF2E1A2C`), gold `AutoAwesome` iconography, and Serif typography.
    - Adjusted `LazyColumn` vertical spacing to 24.dp and updated content padding for better layout consistency.
- **Vanity Screen Refactoring**:
    - Removed the "Glow Sync Hub" surface and logic from `VanityLandingScreen.kt`.
    - Eliminated the redundant "Glow Sync" icon button from the vanity screen's action row.
